chore: rename oauth2client folder
