+++ jarvis_sdk_node/src/sdk/model/index.ts
@@ -1,7 +1,10 @@
-export * from './digitaltwin';
-export * from './token';
-export * from './property';
+export * from './authorization_decisions';
 export * from './config';
 export * from './consent_challenge';
-export * from './import_digitaltwin';
+export * from './digitaltwin';
 export * from './hash_algorithm';
+export * from './import_digitaltwin';
+export * from './ingest/ingest_response';
+export * from './invitation';
+export * from './property';
+export * from './token';
